requirements.txt
# BIS RAG - Runtime dependencies (verified working versions)
# Install: uv pip install -r requirements.txt

torch==2.6.0
sentence-transformers==2.7.0
faiss-cpu==1.13.2
rank-bm25==0.2.2
FlagEmbedding==1.4.0
transformers==4.57.6
datasets==2.19.2
pyarrow==16.1.0
pandas==2.2.2
numpy>=1.26.0
tqdm>=4.0.0
